api/requirements.txt
Flask~=2.3.2
flask-restx~=1.1.0
Flask-Cors~=3.0.10
SQLAlchemy~=2.0.12
Flask-SQLAlchemy~=3.0.3
psycopg2_binary~=2.9.6
marshmallow~=3.19.0
PyJWT~=2.6.0
